services-to-do/cellseg-segment-dtma/requirements.txt
cellmaps-sdk[3_7]
absl-py==0.7.1
albumentations==0.3.1
astor==0.7.1
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cloudpickle==1.0.0
colorama==0.4.5
cycler==0.10.0
cython==0.29.13
dask==1.2.2
decorator==4.4.0
defusedxml==0.6.0
docopt==0.6.2
entrypoints==0.3
gast==0.2.2
google-pasta==0.1.7
grpcio==1.16.1
h5py==2.8.0
idna==2.8
ijroi==0.4.0
imageio==2.5.0
imgaug==0.2.5
jedi==0.13.3
jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.2
keras==2.2.4
keras-applications==1.0.7
keras-preprocessing==1.0.9
kiwisolver==1.1.0
markdown==3.1.1
markupsafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
mock==3.0.5
nest-asyncio==1.5.5
networkx==2.3
numpy==1.16.0
olefile==0.46
opencv-python-headless==4.6.0.66
pandas==0.25.0
pandocfilters==1.4.2
parso==0.4.0
patsy==0.5.1
pbr==5.1.3
pexpect==4.7.0
pickleshare==0.7.5
pillow==6.1.0
pims==0.4.1
pipreqs==0.4.9
prompt-toolkit==3.0.30
protobuf==3.7.1
ptyprocess==0.6.0
pycparser==2.19
pydub==0.23.1
pygments==2.4.0
pyparsing==2.4.0
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.1
pywavelets==1.0.3
pyyaml==5.1.2
pyzmq==18.0.2
read-roi==1.5.2
requests==2.22.0
scikit-image==0.15.0
scikit-learn==0.20.1
scipy==1.1.0
seaborn==0.9.0
send2trash==1.5.0
six==1.12.0
slicerator==1.0.0
statsmodels==0.10.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tifffile==0.15.1
toolz==0.10.0
tornado==6.0.3
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
werkzeug==0.15.3
wrapt==1.11.2
yarg==0.1.9
